requirements/base.txt
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Django==5.2.1
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
Flask==3.1.1
Flask-OAuth2-Login==0.0.9
Flask-SSLify==0.1.5
idna==3.10
inflection==0.5.1
inventory==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
